flask_app/requirements.txt
Flask==1.1.2
numpy==1.16.6
tensorflow==2.2.0
Pillow==7.0.0
Keras==2.3.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0

